Fix Variant to string constructor ambiguity in macos/ios builds

+++ src/ridge_impl/ridge_based_object.h
@@ -15,16 +15,16 @@ class RidgeBased {
   void print_biomes() {
     print("\n");
     for (RidgeGroup& group : _mountain_groups) {
-      print("Mountain group of size ", group.meshes().size());
+      print("Mountain group of size ", (uint64_t)group.meshes().size());
     }
     for (RidgeGroup& group : _water_groups) {
-      print("Water group of size ", group.meshes().size());
+      print("Water group of size ", (uint64_t)group.meshes().size());
     }
     for (RidgeGroup& group : _hill_groups) {
-      print("Hill group of size ", group.meshes().size());
+      print("Hill group of size ", (uint64_t)group.meshes().size());
     }
     for (RidgeGroup& group : _plain_groups) {
-      print("Plain group of size ", group.meshes().size());
+      print("Plain group of size ", (uint64_t)group.meshes().size());
     }
   }
